fix: collection-list pin icon not changed
issue: https://github.com/toeverything/AFFiNE/issues/3498
